Moved security token methods from VVRestAPI.js into common.js to allow httphelper to refresh auth tokens when they're about to expire. Altered acquireSecurityToken method to return the session token instead of a new vvClient object. Modified getVaultApi and getValutApiFromJwt methods to be responsible for instantiating returning a new vvClient object.
